+++ components/cli/include/cli.h
@@ -1,9 +1,11 @@
 #ifndef CLI_H
 #define CLI_H
 
+#include <stdint.h>
+
 // Command buffer max size
 #define CMD_BUF_SIZE 64
 
-void cli_init(int *idx, char *cmd_buff);
+void cli_init(uint8_t *idx, char *cmd_buff);
 
 #endif
